pyproject.toml
[project]
name = "master-thesis"
version = "0.1.0"
description = "Master Thesis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "spflow>=1.1.0",
    "torch>=2.11.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
